requirements.txt
telethon
aiohttp
redis
GitPython